requirements.txt
termcolor==2.4.0
torch==2.4.0
torchvision==0.19.0
wandb==0.17.5
numpy==1.26.4
pillow==10.4.0
matplotlib==3.9.1
tqdm==4.66.4
scikit-learn==1.5.2
opencv-python==4.10.0.84
gdown==5.2.0